+++ uot-server/headers/players.h
@@ -8,18 +8,18 @@
 #include "galaxy.h"
 #include "player.h"
 
-static long player_id = 1;
+static unsigned int player_id = 1;
 class PlayersList
 {
    private:
-    std::map<long, std::shared_ptr<Player>> players = {};
-    std::map<long, std::string> players_net_names = {};
-    std::map<std::string, long> players_net_names_rev = {};
+    std::map<unsigned int, std::shared_ptr<Player>> players = {};
+    std::map<unsigned int, std::string> players_net_names = {};
+    std::map<std::string, unsigned int> players_net_names_rev = {};
 
    public:
     std::shared_ptr<Galaxy> GetStartingGalaxy(std::shared_ptr<Galaxy> wholeGalaxy);
     std::map<Resource, float> GetStartingResources();
-    std::shared_ptr<Colony> GetStartingColony(long player_id, std::shared_ptr<Galaxy> startingGalaxy,
+    std::shared_ptr<Colony> GetStartingColony(unsigned int player_id, std::shared_ptr<Galaxy> startingGalaxy,
                                               std::shared_ptr<Galaxy> wholeGalaxy);
     bool AddPlayer(std::string player_net_name, std::shared_ptr<Galaxy> wholeGalaxy);
     bool HandlePlayerRequests(std::string player_net_name, std::shared_ptr<messageTypes::ActionsPayload>);
